+++ Common/Optimizer/ParameterReport.cs
@@ -20,68 +20,63 @@ using System.Collections.Generic;
 namespace QuantConnect.Optimizer
 {
     /// <summary>
-    /// Sensitivity report for a single optimized parameter, computed from one-dimensional
-    /// cross-sections of the parameter space (slices) where every other parameter is held
-    /// constant.
+    /// Sensitivity report for a single optimized parameter.
     /// </summary>
     public class ParameterReport
     {
         /// <summary>
-        /// Parameter name as defined in the optimization configuration.
+        /// Parameter name.
         /// </summary>
         public string Name { get; set; }
 
         /// <summary>
-        /// Lower bound of the parameter sweep, taken from the optimization configuration.
+        /// Lower bound of the parameter sweep.
         /// </summary>
-        public double SearchedMin { get; set; }
+        public decimal SearchedMin { get; set; }
 
         /// <summary>
-        /// Upper bound of the parameter sweep, taken from the optimization configuration.
+        /// Upper bound of the parameter sweep.
         /// </summary>
-        public double SearchedMax { get; set; }
+        public decimal SearchedMax { get; set; }
 
         /// <summary>
-        /// Sweep step size from the optimization configuration, or null if none was provided.
+        /// Sweep step size; null when not provided in the optimization configuration.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-        public double? Step { get; set; }
+        public decimal? Step { get; set; }
 
         /// <summary>
-        /// Number of distinct values this parameter took across all completed trials.
+        /// Number of distinct values this parameter took across completed backtests.
         /// </summary>
         public int DistinctValueCount { get; set; }
 
         /// <summary>
-        /// Average of (max Sharpe - min Sharpe) across every 1-D slice of this parameter.
+        /// Mean Sharpe range (max - min) across every 1-D slice.
         /// </summary>
-        public double MeanWithinSliceSharpeRange { get; set; }
+        public decimal MeanWithinSliceSharpeRange { get; set; }
 
         /// <summary>
-        /// Maximum of (max Sharpe - min Sharpe) across every 1-D slice of this parameter.
+        /// Maximum Sharpe range (max - min) across every 1-D slice.
         /// </summary>
-        public double MaxWithinSliceSharpeRange { get; set; }
+        public decimal MaxWithinSliceSharpeRange { get; set; }
 
         /// <summary>
-        /// Worst-case Sharpe change between two adjacent grid values of this parameter,
-        /// taken across all slices and scaled by <see cref="Step"/>.
+        /// Worst-case Sharpe change between two adjacent grid values, scaled by <see cref="Step"/>.
         /// </summary>
-        public double MaxAbsDerivativePerStep { get; set; }
+        public decimal MaxAbsDerivativePerStep { get; set; }
 
         /// <summary>
-        /// This parameter's value at the best trial.
+        /// This parameter's value at the best backtest.
         /// </summary>
-        public double BestValue { get; set; }
+        public decimal BestValue { get; set; }
 
         /// <summary>
-        /// True when <see cref="BestValue"/> lies within half a step of the searched min or max,
-        /// signalling that the optimum may lie outside the swept range.
+        /// True when <see cref="BestValue"/> lies within half a step of <see cref="SearchedMin"/> or <see cref="SearchedMax"/>.
         /// </summary>
         public bool BestAtSearchedEdge { get; set; }
 
         /// <summary>
-        /// One-dimensional slices of the parameter space used for the sensitivity analysis.
-        /// Each slice fixes every other parameter and varies only this one.
+        /// One-dimensional slices used for the sensitivity analysis.
         /// </summary>
         public IReadOnlyList<SliceFit> Slices { get; set; }
     }
